Squashing the commit messages together into one.
This change introduces a new role to the fold. This role is designed to replace the Cohga Weve shell script which installs Weave. A shell script isn't ideal and indeed it's actually very complex in this case, require several files and silly, complicated sed commands to introduce what is essentially: templating.