Make rds can be shared for more than 1 role_type
Currently for each role_type we create one RDS.

This change try to introduce a variable rds_application which default to be a
role_type so we no longer be limited by one role_type with one RDS. We can
override the rds_application at inventory to create RDS and then in network
stack modify the 'Application' tag to match with rds_application.

This is for ACT2 use case when we want to use one RDS for both certificate-api
and certificate-db as there is no requirements to split them and it is cheaper
to have only one.