allow-multiple-rds-engine-type-for-onerole
In XVT Cloud we want to allow in a 'role_type' (such as a web platform stack)
can use different RDS engine type (such as postgres or mariadb, etc).

This changes allows us to pick up the correct SG (as different rds have
differrent service ports). Also it allows us to pick up the Public tag to cater
for the case if clients want to open their DB to the public. Crazy but some
wants that.

THIS IS BREAKING CHANGE. Thus the tag major version will be bumped.